backend/requirements.txt
fastapi
uvicorn
openai
PyMuPDF
python-dotenv
requests
beautifulsoup4
